Fix major memory leak after referencing crewmembers
After referencing a CrewMember, it will be stored in that server's
ObjectStore.  The KRPC CrewMember object has a reference to the KSP
ProtoCrewMember associated with the kerbal.  But this object has a
reference to the part where they are sitting in, which in turn has a
reference to the vessel, which in turn has references to all of its
parts.  That means that after loading a quicksave, if you had ever
referenced a crew member through KRPC, you will leak the entire
loaded vessel hierarchy each time a quicksave is loaded.