Fix decoupler crashes with stack separators
This crash seems to stem from the fact that there are two vessels made
by stack separators: the main separated vessel and the stack seperator
debris. Based on some cursory testing, this seems to select the
separated vessel and not the debris. There may be some edge cases here,
but they haven't cropped up in my testing as of yet.